requirements.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
GoogleNews==1.6.6
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_widgets==3.0.15
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==2.3.4
nupy==0.1.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.4
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
transformers==4.44.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.14
widgetsnbextension==4.0.14
nbformat
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyPDF2
openai
langchain